STY: Quiet some compiler warnings.

+++ src/onedtotwod.f
@@ -3,7 +3,7 @@
       real a(nlat,nlat,nt),b(nlat,nlat,nt)
       complex dataspec(nmdim,nt)
 
-      ntrunc = -1.5 + 0.5*sqrt(9.-8.*(1.-float(nmdim)))
+      ntrunc = int(-1.5 + 0.5*sqrt(9.-8.*(1.-float(nmdim))))
 
       scale = 0.5
